Split up storage tests.
