+++ packages/service-core-tests/src/tests/util.ts
@@ -1,0 +1,3 @@
+import { test_utils } from '../index.js';
+
+export const TEST_TABLE = test_utils.makeTestTable('test', ['id']);
